[Res Explorer] Fix create 'X Resource File' action behavior
Initially, I simply tried to re-use the same 'CreateResourceFileAction's
used in the project tree when right-clicking resource folders. That was
a mistake. The behavior of the CreateResourceFileDialog changes too much
when the project tree is in Android View or Project View. While
sometimes useful, it results in a very confusing experience from the
ResourceManager.

The difference, is that from the project view the user can be very
specific in what resource folder to create the resource which implies
information like qualifiers and source set (E.g: right clicking on
../src/debug/res2/drawable-xhdpi). And using the same action from the
'unified' Android View allows the user to choose that information.

As such, this CL reverts many changes from
I589a89954aa0f3f57d6dc995ec49a100ef3beae2.

The ResourceManager now instead calls the dialog trhough
CreateResourceFileAction.createFileResource wrapped in AnAction. So it
should behave very similarly as triggering the action from the Project
tree in the 'Android' view.

Bug: 138257967
Test: n/a
Change-Id: Ieb6da24a91f37812077353e6ad21ede79c7ba24d