[Kotlin Dsl] support dependencies with blocks.
This change adds support to dependencies having blocks (ex : for exclude
statements). When parsing KtCalExpressions, we need to check if a method
has a blockExpression whether it should be parsed as a block (ex :
buildType) or not.
The change also provides support to dslLiterals having a configBlock at
both the parser and writer stages.

Bug: N/A
Test: testFiles added for existing tests, more tests should be added in
future CL for dslMethodCalls with closureBlocks.

Change-Id: I2b164cba8c4f711da16ae3b0a1117977d5bc040e