Improve error handling for Compose preview
This makes sure that the preview is not refreshed is the project did not
compile so the user can keep the last good render.
We also display a notification when the build failed so the user knows
that a build is needed.

Test: Manually tested. In an already compiled project, do a clean, close
and reopen the preview. Also, introduce syntax errors and check the
preview still works when you press refresh.
Fixes: 139519842
Change-Id: Ia80baf9ea104f88faa23d1b664a65db2852a7c60