Save Gradle properties lexicographically
We are currently storing them in the order they are enumerated by the
underlying Hashtable, causing the generated files to not have an
specific order or even the same order when the properties are equal.

After this CL the properties are sorted lexicographically before saving
to a file. In this way the files are stable and easier to read.

Bug: 135102054
Test: GradlePropertiesTest#testGetSortedProperties
Change-Id: I2e52f501cde06fc272aea83bdab37624c280405b