Remove redundant indices in UnifiedEventsTable
Previously we create three sets of indices. Upon inspecting the execution
plans of all the queries we do, have a single index actually works for all
of them. Exception is queries that look for events matching a specify command id,
but those are infrequently and runs asynchronously, so it is okay not having a
separate index just for that use case.

Bug: 139497901
Test: Existing
Change-Id: I03eab05fddfc739e056728f88ad4bc7c4d4edb2e