[KTS Writer] provide support for extra properties and vriables.
This CL adds the possibility of applying changes to properties and
variables, as well as creating them from their dslElements inside a KTS
file.
The change also modifies the way a ktCallExpression is deleted from a
kts file : we no longer delete a callExpression' valueArgumentList
unless we need the delete the callExpression itself (by checking it's
dslElement) (for ex: for maps, we have some cases where we don't want to
delete them although they are empty).

Bug: N/A
Test: added
Change-Id: Ia371bda9c16558b5912e70cb6992447051d6063b