Remove duplicated sync errors in Build View.
Each sync error event was sent to sync view manager twice, once in
AbstractSyncMessages::showNotification when the notification is
created, and another one in PostSyncProjectSetup::finishFailedSync
while sending finish build event.

This change removes the first occurence, since the first one was sent
as MessageEvent, where the title is always the same as message body,
and quickfixes in message body are not clickable.

Bug: NA
Test: inspected.
Change-Id: I7b41c4298155e9035891a0e590b1c9961a68296b