Bug fix on generating complex table updates
In the rare case that a user changes a table in such way that the start
version and final version have no columns in common, there is no need to
do the 4 step update. In this case we can simply drop the old table and
recreate it without having to copy any data.

Test: SqlStatementsGeneratorTest
Bug: 136048932
Change-Id: I6bb9f2f5d68eca8ae4ef441518d7914733cc9477