Minor bug fixes related the the profiler import workflows
- Better support re-importing the same file. Previously, Cpu profiler would
fail silently and memory profiler would create a new session for each import
action. This change would select the session associated with the file if it
was previously imported already.

- Prevent the UI from getting stuck in the loading screen when
importing a trace while there is an in-progress session. In that particular
scenario, the UI got into a state where it thinks it is waiting for a preferred
process to spawn due to the auto-profiling feature. This change disables auto-
profiling after a preferred process has been found to prevent the profilers
from continuously looking for that preferred process.

Bug: 139758562
Bug: 139758897
Bug: 139760630
Test: Updated
Change-Id: Ie91b3b5bb4b6351b1a5b5b0e41011ae355823082