Suppress exception with WNA metrics scrolling
Uninitialized metrics tracker has been causing some crashes in Canary 9,
so this CL is just to suppress that. It does not fix the root issue
because I haven't been able to reproduce it yet. This change itself
does not impact metrics collection, because the tracker is already not
getting initialized in some corner case.

Test: N/A
Bug: 140218425
Change-Id: Iac89b77c5c6d46c967b4eadd0e59b3b2915d50ba