ESC in an editor should select the row in Declared Attributes.
Currently ESC in a text editor will cause the attributes panel in
"Declared Attributes" to loose focus. Thus the user has to use the
mouse to regain focus. With this CL focus is going to the table
instead.

This CL is part of making it possible for users to delete a row in
a attributes table.

Bug: 141022637
Test: Existing tests should pass.
Change-Id: If88b7ec7b27b06d46b61024950517f955e3a891e