Auto edit feature should only activate on letters and digits.
Currently we have logic in an attribute table to activate cell
editing when a user starts typing in a selected table row.
Limit this to letters, digits, and alt-down for opening combo
boxes.

This CL is a preparation for letting the user delete a row in a
table by using the delete key. We do not want to activate a cell
editor in this case.

Bug: 141022637
Test: Existing tests should pass.
Change-Id: I6548ffc658fe6c2f515baa2d47164fd6e410c15a