[Dsl,.kts] preserve Psi in complicated gradle names
We use an internal, normalized dotted-name syntax to refer to Dsl
properties, through GradleNameElement objects.  This syntax maps
essentially 1:1 with Groovy, but the Kotlin Dsl needs conversion to
this syntax to handle embedded block creation or lookup, map element
lookups, and the like.  Define a name converter interface, and use it
appropriately within GradleNameElement methods which create a new
instance based on existing Psi content.

This then allows the GradleNameElement to retain a reference to the
underlying Psi of the content which generated it.  Most of the noise
in this is because both the parser and the writer for each Dsl
language must be able to perform this name conversion, because the
writer must be able to commit changes.  This means that a number of
methods must change signature to propagate the parser/writer through
to the point where it can be used in GradleNameElement constructors.

Bug: 130224064
Test: existing
Change-Id: I768ae1c649e6f8af4a5a367441cf0e591d262ae1