Add "context" property to WorkBench
Add a context property to WorkBench, since it can be used in different
contexts. For example, in the split editor we might have a tool window
of a particular work bench being hidden in design mode but shown in split
mode. Adding the context allows us to restore the state corresponding to
the target context.

Bug: 141176929
Test: AttachedToolWindowTest
Change-Id: Ib2c28028c1cc5f013745fd808ecbb05d4406349f