Fix rendering issues with EnergyProfiler
- Previously for stacked line series in LineChart, we update the underlying
data series in place to accumlate the stacked values. It stopped working due
to a change where we cache the underlying data series to prevent hitting the
datastore too frequently, and the cached series was being modified and reused
across multiple UI frames. This change prevents LineChart from updating the
underlying dataseries for stacked configs. To test this, the DefaultDataSeries
used for tests have been updating to return the same SeriesData instances instead
of creating new ones, which allows us to verify that they are not being modified
in stacked config scenarios.

- Also fixed a bug in Energy Profiler L2 where the Y range interpolation only
looks at the values from each individual usage category instead of the total.
This prevented the LineChart from using the correct Y max value to render the
axis and data.

Bug: 140086078
Test: Added LineChartTest#testStackedConfigsDoNotModifyUnderlyingDataSeries
Change-Id: I00f9d39d7962aee62828af31d1fa85b7d866cea9