[Res Picker] Set proper order of resource tabs
The order of the resources tab should match the order in the Resource
Maanger as much as possible (i.e: resources not supported in the
Resource Manager should be appended at the end at any order).

The initial selected resource tab should either match the one of the
initially selected resource, or the preferred resource type given to the
resource picker.

Bug: 140588703
Test: added tests to ResourceExplorerDialogTest
Change-Id: I4a1f7b1c7d1c66cabe031c25d3a265056bec2232