Add attributes for missing constraint tags.
Refactor the properties panel for Motion layout such that we
track the current selection by: MTag[] and selection type instead
of the current MotionSceneTag.

Make all motion selection events go through a queue in the
properties panel. This was accidentally disabled.

Add top level panels for better identification of what in the
motion scene file the attributes are displayed for.

Bug: 118388686
Test: Existing tests should pass.
Change-Id: I3ac7d0907aff09d824c98d3114b1f85314ac6ec0