Use EnvironmentUtil instead of TeminalUtil
... when reading environment variables.

TerminalUtil.java was created with the sole purpouse to get environment
variables from a terminal, but it is causing some Android Studio freezes
when the created processes have problems. It was based on
EnvironmentUtil but some of the time out handling was not done in the
same way, causing a different behaviour.

This CL removes TerminalUtil an duses EnviromentUtil to have a
consistent behaviour.

Test: existing, manually checked on Windows, Linux and Mac that
JAVA_HOME was read correctly.
Bug: 140406172
Change-Id: I214dc709b970ce74459489e03660cce9a9a83fd5