In the properties table: toggle a group only when icon is clicked.
Currently it is not possible to select a row in a table by clicking
on the name of a property that is a group (flag or property with
source locations in the inspector).

This CL allows this by making the user press on the group icon to
open/close the group. Clicking on the rest of the name column will
cause a row selection.

Bug: 123643052
Test: N/A
Change-Id: Ifd5ca6384d84ad35a76f9ad4cfb4ea46fd2cdc81