Experimental: AppInspectionDiscovery
It creates new module app-inspectionAndroid that adds
LaunchTaskContibutor, that does nothing by default.

It also adds a new studio flag to turn on sqlite inspector,
the same enables AppInspectionDiscovery, mechanism
that abstracts away device discovery and attachment to it from the clients.

Code in this CL is be gated by the flag and is going though major rewrite.

bug: 143628758
Test: flagged experiment, should go through major rewrite before publish
Change-Id: Iadc7a4a97144f46c66780380a4518166279423ee