Set default properties when switching workbench context
When we were setting a workbench context for the very first time we open
studio and then tried to update the model, we ended up defaulting the tool
window states to false, which would cause all of them to be minimized on
the top-right corner of the workbench.

That was exposed by some UI tests that would fail depending on the order
they ran. The following steps are an example of what was happening in a
failing scenario:

1) Open a layout file
2) Set workbench context to the current mode after initializing the model
3) Call Workbench#init, which will set the default properties correctly
   since the context was already set
4) Change the editor mode, which will trigger a listener to update the
   context to the new mode and update the model.
5) When updating the model, we'd check the default properties of the current
   mode and use the default value of PropertiesComponent#getBoolean because
   the values are not set.
6) Tests would fail because they were trying to access elements inside the
   tool windows that were minimized. For the end user, the tool windows
   would be unexpectedly collapsed, so they needed to manually open them.

This CL fixes this issue by making sure we set the default properties for a
context, if they're not already set.

Bug: 141549717
Test: Added unit test. UI tests should pass consistently now.
Change-Id: Ia641326360096b32ed00d1110bfccf431d94b0d9
(cherry picked from commit eb37e93210628b9fe8cef76bdd196070c2f089e8)