[Asset Studio] Provider class for new material icons in IconPickerDialog
Provider class that call MaterialIconsLoader asynchronously and handles
getting the metadata file for it. In general this class will handle the
logic to provide material icons to the ui (IconPickerDialog).

Also added necessary Gson parser for the icons' metadata.

Moved and added some useful interfaces for testing. As well as some
utility methods.

Bug: 141629663
Test: MaterialVdIconsProviderTest, MaterialIconsMetadataTest
Change-Id: I54d9fa5209db5ce1ed364f94ab4e8994d23797c0