Small LayoutEditorFileType refactoring to allow for Compose
LayoutEditorFileType currently forces a palette to be available. We do
not want (yet) a palette in compose but we want to be able to re-use
the stats reporting that come with the LayoutEditorType.
Also move the logic of calculating the file name to the palette model
since the type should not necessarily know how the name is created.

Bug: 141857481
Test: Existing tests pass
Change-Id: I6968af9f6ea8bfbf825d4cbc058206fc5deb8395