[Dsl] semantic name translations in NameConverter interface
"semantic" here meaning respecting the semantics of the Project object
that our Dsl is implicitly modelling, as opposed to the syntactic
translation performed by psiToName() / convertReferenceText(): the
semantic element is that these name translations depend on the
context; the Dsl element in whose scope the name is being resolved.

We define modelNameForParent(), which for a given external name and a
given context (DslElement) provides our internal model name, and
externalNameForParent(), which performs the reverse translation.  The
concrete implementation DslNameConverter classes can perform this
translation in a language-specific way.  In each case the context in
question is the parent element of the element whose name we are
considering.

The language-specific dispatch is built by extending the
GradleDslElement interface to have a method to return an appropriate
map.  Provide a default implementation returning an empty immutable
map, and use this implementation in the GradleDslNameConverter
implementation.

The concept we are trying to represent here, converting a string based
on the current GradleDslElement and the current external language, is
naturally represented by multiple dispatch.  Unfortunately, JVM-based
languages tend not to support multiple dispatch, instead privileging
the receiver; the slightly contorted interface allows emulation of the
dispatch on the name converter class where necessary.  (There may be a
better way)

Things remaining to do:

- there's a clear need for both forward and backward maps.  Update this
  to use a bidirectional map rather than an explicit search

- the Kotlin implementation of modelNameForParent and
  externalNameForParent, and the Groovy implementation of
  externalNameForParent, need to handle dotted names.

Bug: 141842964
Test: exercised in forthcoming CLs
Change-Id: If0cc27ac81a384acb5bdb7ab615b426e3b1cbbea