[Dsl] renaming properties back when writing
When writing our model properties back out, we need to translate
between our model property names and the external name for whichever
configuration language we are writing to.

The principal place where this happens is in each language writer's
maybeUpdateName and in the shared utility maybeTrimForParent: these
routines are explicitly converting from our internal names to an
external format, and so require a name converter to do any
context-specific renaming.  Give each of these routines a
NameConverter argument, and arrange for callers to pass the
appropriate writer class through.

When committing a name change, we construct a psiElement, write it to
the file - and then we update the GradleNameElement from that
newly-created Psi.  That operation mirrors the initial construction
from psi, giving the GradleNameElement an external name, which we must
then convert to a model name at this point, because the element with
its new name is already contained within our Dsl/model structure.
After setting up the new name from the Psi, canonize the name to the
appropriate model name.

Although this now allows us to get the names of external properties
correct when writing, the differences between Kotlin and Groovy syntax
regarding method calls against assignments mean that this is not (yet)
a complete solution even to the simple reading and writing of
properties.

Bug: 130224064
Bug: 141842964
Test: existing (new tests to come)
Change-Id: I730855f65bac1228948bc2c9970b88a6ad8e2203