[Dsl] create consumerProguardFiles as method by default
The information about whether something should be a method or an
assignment will eventually be moved down from the Model layer to the
Dsl/Parser.  However, in the interest of having more correct expected
files in test cases, make this change here and now to mirror the
treatment of the proguardFiles method.

(The method and assignment have different semantics!  the method
appends to an existing list, whereas the assignment assigns.  This
suggests that our future needs to distinguish clearly between these
two operations)

Bug: 141548181
Test: existing tests are made more correct
Change-Id: I89bc727d57fdf9b4a0a4bd6c87cd24f19b4c0600