[Dsl] document the reason for a disabled test
Also document a suggested remedy: if we implement setProguardFiles()
as having assignment semantics (even though it has method form) in
Kotlin, it will not be removed.

The expected form must be listOf<String>(): without the explicit type
decorator, there is insufficient information for what the concrete
type of the empty list should be.

Bug: 72853928
Bug: 141548181
Test: added, though disabled
Change-Id: I6d2c53dd2186e892f4cac77e8a8998f8fbabb028