Minimize tool windows by default in split mode
When opening the split view of split editor for the first time, users
might feel overwhelmed by the amount of tool windows open, giving little
space for the design surface itself. To alleviate that, we minimize the
tool windows by default in split mode.

This CL adds the possibility to do the same for other workbenches as
well, since it might be a common problem considering the amount of tool
windows that can be open at the same time in Studio.

Bug: 142060662
Test: Unit test
Change-Id: I4a512d19c478618c4e27cfeba4fae4d02f8877c3
(cherry picked from commit 8c94d415a5611253a57a2838e11f36dda1ae15d8)