Exclude additional classes (e.g. R.jar) from unit testing classpath
Refactored all places that modified unit testing classpath based on
paths stored in the model and not represented in IntelliJ project
structure, to use the same helper methods. This way we add and remove
matching sets of files based on dependency information read from the
model.

When adding R.jar to the model we discovered the "additional" classes
were not handled correctly by the IDE. This is not a huge deal for R.jar
but the same mechanism is used for kotlinc output, which should be fixed
now.

The most obvious way this can be reproduced is making module A an
androidTestImplementation dependency of module B. When unit-testing B,
"additional classes" from A were available on the classpath. These
classes include R classes and output from kotlinc.

Test: AndroidJunitPatcherWithProjectsTest
Bug: 141372007
Change-Id: I82a8b3afdb402cb431d6456340a93d8d821ae252