[PSD] ProductFlavor dimensions tests for Kotlin
Make the various test cases operate on both the Groovy and the Kotlin
psdSample projects.

In fact testEffectiveDimensions currently tests a mixed project: the
nested1 and nested2 build files are still expressed in Groovy.  There
is no particular harm in this, and the test does not strongly expose
the underlying language of the nested build files -- and there will
also be no harm if those nested build files are rewritten.

As an incidental change, rename the -Kt suffix to -Kotlin in test
names.

Test: more PsProductFlavorTest
Bug: 130224064
Change-Id: I3663bd5e09585910276ec3be3d3d0bdd331e5636