[Dsl,.kts] allow for complicated non-reference DslExpressions
To support some constructs needed in Kotlin build files, we have to
parse postfix expressions `<...>!!` and type-decorated expressions
`<...> as Type`, where we will perform resolution on the bare
expression denoted by `<...>`.  Therefore we have to determine whether
we have a reference or a Kotlin literal by getting the expression
element from that bare expression itself.

Test: upcoming AndroidGradleTest
Bug: 130224064
Change-Id: I4905f3a102de1f34123ba075166dda7c1cf222f1