[Asset Studio] Incrementally load VectorDrawable icons to the dialog
Instead of waiting for all 5,000+ icons to load and then show them in
the ui, load the icons incrementally, by their family/style.

This will result in loading ~1,000 icons at a time and will allow us to
show a usable list of icons to the user nearly as fast as it used to
with material 1 icons (959 icons). Without freezing the ui.

Bug: 141629663
Test: updated existing tests
Change-Id: Iddfeb688a2419575d37ed61d1a92aad5fbb54551