Fix NlPropertyTableTest after changing tab order in designer
The tab order in the layout designer was changed in Change Ie6c7f9
such that the user can navigate between the properties panel, the
design surface, palette, and component tree. Before that change
focus stayed in the properties panel and this test was tabbing
from the last component to the top of the properties panel.
Changed the test not to do that.

Bug: 142354157
Test: This fixes a UI test failure.
Change-Id: Ia022fdfe8267a9bec2004a30a9e64ed71dc24f4d