Fix handling of "Fatal IDE Errors" in UI tests
Errors may not appear in the result of MessagePool.getFatalErrors until
a second or more later due to aggregation logic in MessagePool. Thus,
errors from the end of one test may not appear in time for the post-test
check, and end up appearing in the subsequent test.

This change waits for a dummy error to propagate through MessagePool
before asking for the list of fatal errors.

Fixes: 142425546
Change-Id: I13ed3a429c429f517a9c2cee2d85a5a91dee2600