[Dsl,.kts] apply plugin tests
This reveals a problem with our current implementation of apply
plugin; log a bug for it and annotate source and test cases.  It also
reveals one use of an interface apparently retained for compatibility
with something that I can't find any other references to; log that as
a bug too.

Test: strengthen existing ApplyPlugin tests
Test: run existing tests on .gradle.kts build files
Bug: 130224064
Bug: 142114586
Change-Id: I22fb5a39f2866a9b5b588d5cae5bed09b629706b