Use cached Configuration device explicitly when possible
Configuration#getDevice() is a @Slow method because it can potentially call
Configuration#computeBestDevice, which is in fact a slow method. The device
is computed by ConfigurationManager#getConfiguration, when it's called from
NlModel#create. Calls to getDevice() after that in the design tools context
will return the cached value, which has a trivial cost.

This CL makes the callers of getDevice() that are expected to use the cached
value to explicitly get the current device without having the chance to call
computeBestDevice().

Bug: 128894267
Fixes: 142532429
Test: UI tests, OrientationMenuActionTest
Change-Id: Ie3374c07adf7b31a2030a020f004a32bd07c8496