[Dsl] delete "release" and "debug" BuildType blocks if empty
The release and debug BuildTypes are automatically created by Gradle.
In particular for Kotlin, this means that empty blocks for those build
types are side-effect free and can be deleted.  (They can for Groovy
too; because of Groovy's automatic build type creation, so can other
build types, but in the interest of consistency delete just the two
special build types for now.

Bug: none
Test: added BuildTypeModel tests
Change-Id: I6562fc3b583e5c1774c591a5965aa9c6a0af5d2b