Fix GradleOutputParser to use BuildEventDispatcher
With intellij 2019.2, BuildOuputReader is replaced by
BuildEventDispatcher.
This change refactored Studio's handling of build output
parsers and make it use the new BuildEventDispatcher APIs.

With the new API, underlying output parsers are provided by
GradleOutputParserProvider extension, and there is no API
to retrieve them. So this cl created a project service
BuildOutputParserManager to proide the parser instances
and this ensures that the same output instances are used
for each project.

With linked build, there will be multiple build tasks per project
thus requries parsers per task. But since gradle build in general
don't support linked project, this is omitted in this cl to avoid
unneeded complexity. Current implementation can be updated easily
to support linked build when we do that in the future.

Bug: NA
Test: Covered by BuildOutputParsersIntegrationTest
Change-Id: I3792a6b8f87f9467c406e70d8ad442a2063c56dd