Fix draw order.
Previously we placed a view into a level as follows:
- The primary level of a view is equal to its depth in the view tree
- If there are multiple overlapping views in the same level, the level
is split into sublevels such that earlier views in the level are beneath
later views.

This prevents views from being drawn overlapping in the same plane in 3d
view, but doesn't result in things being drawn in the correct order. For
example, if we have
V1
- V2
  - V3
- V4
where V2 and V4 are children of V, and V4 overlaps with V3 and V2, V3
will incorrectly be drawn on top of V4.
In this CL the algorithm is changed:
- Each view is placed in the first level with no views that overlap it.

In addition to being simpler, I believe it's also more correct.

Test: updated DeviceViewPanelModelTest.testOverlappingRects
Bug: none
Change-Id: I7b98ece9262c8d05a973d39f781b01fc126b6831