Remove ProcessModel reference from ATraceFrameManager.
With this CL, we don't hold any extra reference to the Trebuchet model
after the parsing is done.

While this disallows on-the-fly FramesList building, all the entry
points to buildFramesList currently pass default filters that matches
only one of the two pre-computed options. When/if we have an extended
use case, we revisit how we store them.

As of this, by releasing the Trebuchet models, we hold 24 MB less when
parsing a 2.8 MB ATrace/Perfetto capture.

Test: existing + manual profile
Bug: 143617928
Change-Id: I8e9feb92f2cd2a6099fa14c4691171264eb56d59