[Dsl,.kts] better defaultConfig handling
The defaultConfig block in the android model is essentially the same
as a productFlavor, except that it is not a NamedObject (and the
android block does not act as a NamedObjectContainer).  In immediately
practical terms, this means that when writing out a defaultConfig
block we must not attempt to create it, or even look it up by name.

Make a distinction in the Dsl class hierarchy between actual
ProductFlavor Dsl elements and the special DefaultConfig Dsl
element.  Only the ProductFlavor Dsl elements are NamedObjectElements;
the DefaultConfig Dsl element is not.

Bug: 130224064
Bug: 142114586
Test: stronger addAndApplyDefaultConfigBlock test (and run for .kts)
Change-Id: Ia7256a26da7c5ca3638ef655a7c1a5796e977dcd