For a build from run config, do refresh of VFS post build synchronously.
After build is complete, PostProjectBuildTasksExecutor will refresh the
VFS. When the build is part of run configuration, we want to refresh
synchronously.
E.g. when running unit tests, the test runner will add
paths of compiled classes to unit test runtime classpath. A race
condition might occur when refreshing the VFS async-ly causing the
compiled classes not to be found when adding the classpath, causing it
to be omitted, which in turn causes the test to fail to run.

Bug: 139793893
Test: Manual
Change-Id: I8c5166625b53c8d174232a92dfe0a499db0dbe82