Find a component from constraint set for KeyCycle & KeyAttribute
Pass an component to the new custom attribute dialog such that we
can provide completions for these custom attributes.
First attempt to select a component in the start constraintSet of
the transition being edited. If not found pass an arbitrary
component.

Bug: BurnList
Test: N/A
Change-Id: I0a9b7e0944c7f8bfd4b610d0b50a1fad56953142
(cherry picked from commit 222153800206f95493684833d8bfe3cad8906f66)