Update the MTag after making an update.
Update the attributes in a MTag after making an update to the
underlying XmlTag. The UI may cause a repaint that would be using
the values of the MTag before the update thus making the value
flicker in the properties panel.

Bug: 142431987
Test: N/A
Change-Id: I229f824457707b5ae3b6de9d856ebc1df19a2409
(cherry picked from commit a370baa27e565cc0275361cb13390a85f1a7c608)