Change field name from "isStatic" to "memberAccess"
This CL revisit the logic in whether a class or instance
could access its static/non-static members:

A class could only access its static members: String, Map...
An instance could access both static and non static members.
Only non-static methods are allowed as a getter or setters
to a field e.g. `getField` can not be static as a getter
to `field`.

Fixes: 141383218
Test: DataBindingExprReferenceContributorTest
Change-Id: Ib512e4205e3366d23670a7ce1f7c1acbf389b727