Set max zoom to fit level for Compose preview
In compose, we usually show several small components. The existing zoom
to fit behaviour would make the elements to look extremely large with
little benefit for the user.
After this change, Compose preview limits the zoom to fit to 200% (the
user can still set any zoom level they want).
We also force a zoom to fit when the user changes the "Show decorations"
setting since it changes significantly the size of the rendered canvas.
By automatically calling zoomToFit we make sure the user sees all the
components.

Fixes: 143606329
Test: Existing tests pass
Change-Id: I6760470a72b3c86a20c1b76f225fb951c28c7b04