Extract template tests to their own module
These account for maybe a quarter of the execution time of
intellij.android.core.tests, which is typically the bottleneck for
continuous-build latency. By moving them to a separate module with fewer
dependencies, they can start earlier in the test run.

After this change, 10 shards will be allocated to template tests and 40
to the rest of core tests. This split roughly equalizes the lengths of
the longest shard in each target.

Eventually the associated production code should probably be moved into
its own module as well.

Bug: 67313506
Change-Id: Ia244eae8670d320e9d330b39366c886f02ef4d76