143868566: Missing layout classes are not shown as unresolved in the editor
First, this CL adds an inspection which validates that layout tags in
XML files correspond to resolvable classes, and if not, highlights these
in red as is done for various other resolve validation checks.

Second, this CL updates the Maven library import facility to offer a
quickfix for these unresolved symbols when they correspond to a known
library, such as ConstraintLayout or RecyclerView.

Test: Unit test included
Fixes: 143868566
Fixes: 143868567
Change-Id: I89c62019e7214da8f7437122cfbd452169904380