Validate default activity using AndroidManifestIndex
If StudioFlags.DEFAULT_ACTIVITY_LOCATOR_STRATEGY
is set to INDEX, then DefaultActivityLocator should use the
AndroidManifestIndex to gather a list of activities accessible from a
particular module when executing a run configuration. We do this to
avoid the staleness and performance issues that come with using
MergedManifestManager. Currently, we only leverage the index when we
go to launch the default activity, but not during run config validation,
which still relies on MergedManifestManager. But this doesn't make sense
because we always run the validation logic before executing a run
config, which means we hit the MergedManifestManager issues either way.

This CL updates DefaultActivityLocator#validate to use
AndroidManifestIndex as well if the strategy is set to INDEX so that
we can avoid MergedManifestManager as we originally intended. The
validate() method is called very frequently, especially on Macbooks,
where it's used to update touchbar icons. So in order to avoid a
drastic increase in our metrics log storage, this CL moves the
logic responsible for recording latencies down into
doGetActivitiesFromManifestIndex. This means we will only record
latencies for cache misses, which is something we will now have
to take into consideration any time we're comparing these latencies
with metrics for the other strategies.

Bug: 143693655
Test: DefaultActivityLocatorTest
Change-Id: I12e1676a15287b7f0d1f5e461d0452b4dabed53a