[Dsl,.gr] add map arguments at the end
A Groovy GrCommandArgumentList is structurally different from a normal
one, in that its last child is an argument, rather than the `)` to
close a method call's argument list.  However, the addNamedArgument
method does not take account of this distinction, and inserts new
arguments before the last argument of a GrCommandArgumentList rather
than at the end.

Work around by calling addAfter() on GrCommandArgumentLists, which
does do what we need.

Fixes: 144075798
Test: existing tests updated
Change-Id: Ie75ee47d6c86b9544d3c300115543f63050f8e12